Fix deploy & Linter error
