+++ .eslintrc.json
@@ -14,7 +14,7 @@
   },
   "extends": ["airbnb", "plugin:react/recommended"],
   "plugins": ["react"],
-  "rules": {
+  "rules": { 
     "jsx-a11y/click-events-have-key-events": "off",
     "jsx-a11y/no-noninteractive-element-to-interactive-role": "off",
     "jsx-a11y/no-noninteractive-element-interactions": "off",
@@ -22,6 +22,7 @@
     "react/react-in-jsx-scope": "off",
     "import/no-unresolved": "off",
     "no-shadow": "off"
+    
   },
 
   "ignorePatterns": ["dist/", "build/"]
